Dev imran

+++ index.js
@@ -82,12 +82,12 @@ async function run() {
     });
 
     // Single restaurant data API
-    app.get('/singleRestaurant/:id', async(req, res) =>{
+    app.get("/singleRestaurant/:id", async (req, res) => {
       const id = req.params.id;
-      const query = {_id: new ObjectId(id)}
+      const query = { _id: new ObjectId(id) };
       const result = await restaurantCollection.findOne(query);
-      res.send(result)
-    })    
+      res.send(result);
+    });
 
     // business apis
     app.post("/business", verifyJwt, async (req, res) => {
@@ -146,24 +146,36 @@ async function run() {
       const data = req.body;
       const filter = { email: data?.email };
       const findUserusers = await usersCollection.findOne(filter);
-      const updateDoc = {
-        $set: {
-          ...findUserusers,
-          role: "partner",
-        },
-      };
-      const result1 = await usersCollection.updateOne(filter, updateDoc);
-      const result2 = await partnerCollection.insertOne(data);
-      res.send({ result1, result2 });
-      const isExistInBusiness = await businessCollection.findOne(filter);
-      const isExistInRider = await riderCollection.findOne(filter);
-      if (isExistInBusiness) {
-        const result3 = await businessCollection.deleteOne(filter);
-        res.send(result3);
-      }
-      if (isExistInRider) {
-        const result4 = await riderCollection.deleteOne(filter);
-        res.send(result4);
+      if (data.outletName) {
+        const updateDoc = {
+          $set: {
+            ...findUserusers,
+            role: "partner",
+          },
+        };
+        const result1 = await usersCollection.updateOne(filter, updateDoc);
+        const result2 = await partnerCollection.insertOne(data);
+        res.send({ result1, result2 });
+        const isExistInBusiness = await businessCollection.findOne(filter);
+        const isExistInRider = await riderCollection.findOne(filter);
+        if (isExistInBusiness) {
+          const result3 = await businessCollection.deleteOne(filter);
+          res.send(result3);
+        }
+        if (isExistInRider) {
+          const result4 = await riderCollection.deleteOne(filter);
+          res.send(result4);
+        }
+      } else {
+        const partnersData = await partnerCollection.findOne(filter);
+
+        // Add the entire data object to the menu array
+        if (partnersData) {
+          const updatedMenu = [...(partnersData.menu || []), data];
+          await partnerCollection.updateOne(filter, {
+            $set: { menu: updatedMenu },
+          });
+        }
       }
     });
